+++ src/main/java/me/yleoft/zAPI/utils/EnchantmentUtils.java
@@ -1,0 +1,11 @@
+package me.yleoft.zAPI.utils;
+
+import org.bukkit.enchantments.Enchantment;
+
+public abstract class EnchantmentUtils {
+
+    public static Enchantment getEnchantment(String enchantmentName) {
+        return Enchantment.getByName(enchantmentName);
+    }
+
+}